refactor: move getPowerAnalysis API helper up to experiment-utils.ts
